Flatten the heading levels
Given that the bulk of the content (the data sections) are on the same level, I think it makes sense to keep all the headings on the same level, to simplify the table of contents and structure.

This also lets us increase the font size of those headings, and the table headings (captions), which otherwise are quite small (same size as body text).